+++ web_backend/app/api/v1/views/vitals.py
@@ -2,15 +2,19 @@
 """ objects that handle all default RestFul API actions for Vital Signs """
 from flasgger.utils import swag_from
 from flask import abort, jsonify, make_response, request
+from flask_login import current_user, login_required
 
 from models.notes.vitals import VitalSign
 from storage import storage
-from api.v1.views import app_views
+from web_backend.app.roles import RBAC
+from . import api_views
 
 
-@app_views.route('/patients/<int:pid>/vitals', methods=['GET'],
+@api_views.route('/patients/<int:pid>/vitals', methods=['GET'],
                  strict_slashes=False)
-@swag_from('documentation/patient/patient_id/get_vitals.yml')
+@swag_from('documentation/patients/get_vitals.yml')
+@RBAC.allow(['nurse'], methods=['GET'])
+@login_required
 def get_vitals(pid):
     """
     Retrieves the all VitalSign object for a specific patient
@@ -23,11 +27,13 @@ def get_vitals(pid):
     return jsonify(vitals)
 
 
-@app_views.route('/patients/<int:pid>/vitals/<vitals_id>', methods=['GET'],
+@api_views.route('/patients/<int:pid>/vitals/<vitals_id>', methods=['GET'],
                  defaults={'consult_id': None}, strict_slashes=False)
-@app_views.route('/patients/<int:pid>/consultations/<consult_id>/vitals',
+@api_views.route('/patients/<int:pid>/consultations/<consult_id>/vitals',
                  methods=['GET'], strict_slashes=False)
-@swag_from('documentation/patients/patient_id/consultations/consult_id/vitals/get_vital.yml')
+@swag_from('documentation/patients/vitals/get_vital.yml')
+@RBAC.allow(['nurse'], methods=['GET'])
+@login_required
 def get_vital(pid, consult_id, vitals_id=None):
     """
     Retrieves the vitals for a consultation
@@ -54,12 +60,14 @@ def get_vital(pid, consult_id, vitals_id=None):
     return jsonify(vitals.to_dict())
 
 
-@app_views.route('/patients/<int:pid>/vitals/<vitals_id>', methods=['DELETE'],
+@api_views.route('/patients/<int:pid>/vitals/<vitals_id>', methods=['DELETE'],
                  defaults={'consult_id': None}, strict_slashes=False)
-@app_views.route('/patients/<int:pid>/consultations/<consult_id>/vitals',
+@api_views.route('/patients/<int:pid>/consultations/<consult_id>/vitals',
                  methods=['DELETE'], strict_slashes=False)
-@swag_from('documentation/patients/patient_id/consultation/consult_id/vitals/delete_vitals.yml',
+@swag_from('documentation/patients/vitals/delete_vitals.yml',
            methods=['DELETE'])
+@RBAC.allow(['nurse'], methods=['DELETE'])
+@login_required
 def delete_vitals(pid, consult_id, vitals_id=None):
     """
     Deletes the vitals for a consultation
@@ -83,17 +91,22 @@ def delete_vitals(pid, consult_id, vitals_id=None):
     if not vitals:
         abort(404, description="No such vitals")
 
+    if vitals.created_by != current_user.staff_id:
+        abort(403)
+
     storage.delete(vitals)
     storage.save()
 
     return make_response(jsonify({}), 200)
 
 
-@app_views.route('/patients/<int:pid>/vitals',
+@api_views.route('/patients/<int:pid>/vitals',
                  methods=['POST'], strict_slashes=False)
-@app_views.route('/patients/<int:pid>/<consult_id>/vitals',
+@api_views.route('/patients/<int:pid>/<consult_id>/vitals',
                  methods=['POST'], strict_slashes=False)
-@swag_from('documentation/patient/patient_id/consult_id/post_vitals.yml', methods=['POST'])
+@swag_from('documentation/patients/vitals/post_vitals.yml', methods=['POST'])
+@RBAC.allow(['nurse'], methods=['POST'])
+@login_required
 def post_vitals(pid, consult_id=None):
     """
     Creates a new vitals for a specific consultation
@@ -123,18 +136,20 @@ def post_vitals(pid, consult_id=None):
     data['pid'] = pid
     if consult_id is not None:
         data['consultation_id'] = consult_id
+    data['created_by'] = current_user.staff_id
     vitals = VitalSign(**data)
     vitals.save()
     vitals = storage.get('VitalSign', 'id', vitals.id)
     return make_response(jsonify(vitals.to_dict()), 201)
 
 
-@app_views.route('/patients/<int:pid>/vitals/<vitals_id>', methods=['PUT'],
+@api_views.route('/patients/<int:pid>/vitals/<vitals_id>', methods=['PUT'],
                  defaults={'consult_id': None}, strict_slashes=False)
-@app_views.route('/patients/<int:pid>/<consult_id>/vitals', methods=['PUT'],
+@api_views.route('/patients/<int:pid>/<consult_id>/vitals', methods=['PUT'],
                  strict_slashes=False)
-@swag_from('documentation/patient/patient_id/consult_id/put_vitals.yml',
-           methods=['PUT'])
+@swag_from('documentation/patients/vitals/put_vitals.yml', methods=['PUT'])
+@RBAC.allow(['nurse'], methods=['PUT'])
+@login_required
 def put_vitals(pid, consult_id, vitals_id=None):
     """
     Updates a vitals for a specific consultation
@@ -157,6 +172,9 @@ def put_vitals(pid, consult_id, vitals_id=None):
 
     if not vitals:
         abort(404, description='No such vitals')
+
+    if vitals.created_by != current_user.staff_id:
+        abort(403)
 
     if not request.get_json():
         abort(400, description="Not a JSON")
